missed updating the item constructor to take in a double instead of an int.

+++ wistlish-backend/src/main/java/com/example/wistlish_app/models/Item.java
@@ -30,7 +30,7 @@ public class Item {
 
     // Constructors
     public Item() {}
-    public Item(String name, String itemUrl, String imageUrl, int quantity, int cost, Wishlist list) {
+    public Item(String name, String itemUrl, String imageUrl, int quantity, double cost, Wishlist list) {
         this.name = name;
         this.itemUrl = itemUrl;
         this.imageUrl = imageUrl;
